% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_empty_rows_columns.R
\name{drop_empty_cols}
\alias{drop_empty_cols}
\title{Drop empty columns}
\usage{
drop_empty_cols(mat = NULL, na.rm = TRUE)
}
\arguments{
\item{mat}{A matrix containing interaction data.}

\item{na.rm}{Logical. If TRUE, NA values will be ignored for deciding if a column
should be removed. If FALSE, columns having NA values will never be removed.}
}
\value{
A matrix
}
\description{
Remove columns from an interaction matrix where all values are zero.
}
\examples{
mat <- matrix(c(0,0,0, 1,1,0, 0,0,NA), ncol = 3, byrow = TRUE)
mat
drop_empty_cols(mat)
drop_empty_cols(mat, na.rm = FALSE)
}
