% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_unobs.R
\name{expand_unobs}
\alias{expand_unobs}
\title{Expand interaction dataset to include zeroes (unobserved interactions)}
\usage{
expand_unobs(
  df = NULL,
  plant.var = "Plant",
  animal.var = "Animal",
  int.var = "Visits",
  fill.value = 0
)
}
\arguments{
\item{df}{A data frame with interaction presence or frequency data}

\item{plant.var}{character. Name of the column representing plants.}

\item{animal.var}{character. Name of the column representing animals.}

\item{int.var}{character. Name of the column representing interaction presence
or frequency.}

\item{fill.value}{Value to use in \code{int.var} for unobserved interactions.}
}
\value{
A data frame
}
\description{
Expand interaction dataset to include zeroes (unobserved interactions)
}
\examples{
data(web)
df <- web[sample(1:nrow(web), size = 30), ]
df.expand <- expand_unobs(df)
}
