% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_spec.R
\name{indiv_spec}
\alias{indiv_spec}
\alias{TNW}
\alias{WIC}
\title{Niche width and individual specialisation indices}
\usage{
indiv_spec(net)

TNW(net)

WIC(net, indiv = FALSE)
}
\arguments{
\item{net}{A matrix or data.frame containing interaction data.
Plants in rows, Animals in columns. Numbers need not be integers (i.e. counts),
can be relative abundances or interaction frequencies.}

\item{indiv}{Logical. If TRUE, return individual Shannon diversity values.
If FALSE, return the weighted average of individual Shannon diversity,
weighted by the proportion of interactions represented by each plant
(i.e. the WIC).}
}
\value{
A numeric value, vector, or list, depending on the function
}
\description{
Calculate indices of niche width and individual specialisation,
following Bolnick et al. 2002
(\doi{doi:10.1890/0012-9658(2002)083[2936:MILRS]2.0.CO;2}).
}
\details{
\code{TNW} calculates the Total Niche Width of the population, using Shannon diversity.

\code{WIC} calculates the Within-Individual Component of niche width
as a weighted average of individuals' Shannon diversity
(i.e. the Shannon diversity value of each individual plant is weighted by
the proportion of visits or interactions received by that plant).
\code{WIC} can also return the individual Shannon diversity values, if \code{indiv = TRUE}.

\code{indiv_spec} returns a list with the following components:
\itemize{
\item WIC (Within-Individual Component),
\item Shannon (individual Shannon diversity values),
\item TNW (Total Niche Width), and
\item IndSpec (Individual Specialisation index, calculated as the ratio WIC / TNW).
}
}
\note{
Ideally \code{net} should not contain missing data (NA). If present, they will
be ignored.
}
\examples{
data(web)
net <- long2wide(web)

indiv_spec(net)
WIC(net)
WIC(net, indiv = TRUE)
TNW(net)
}
\references{
Bolnick DI, LH Yang, JA Fordyce, JM Davis, and Svanback R. 2002.
Measuring individual-level resource specialization. Ecology 83: 2936-2941.
}
\seealso{
\code{\link[RInSp:WTdMC]{RInSp::WTdMC()}}
}
