% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2wide.R
\name{long2wide}
\alias{long2wide}
\title{Transform interaction data from long to wide format}
\usage{
long2wide(
  df = NULL,
  plant.var = "Plant",
  animal.var = "Animal",
  int.var = "Visits",
  exclude.noint = TRUE,
  fill.value = NA,
  sort = TRUE
)
}
\arguments{
\item{df}{A data frame with interaction presence or frequency data}

\item{plant.var}{character. Name of the column representing plants.}

\item{animal.var}{character. Name of the column representing animals.}

\item{int.var}{character. Name of the column representing interaction presence
or frequency.}

\item{exclude.noint}{Logical. Exclude plants/animals with no interactions?
Default is TRUE.}

\item{fill.value}{Value to fill empty cells in the matrix. Default is NA.}

\item{sort}{Sort matrix rows and columns by frequency?}
}
\value{
A matrix with plants and animals as row and column names, respectively.
}
\description{
Transform interaction data from long to wide format
}
\examples{
data(web)
long2wide(web)

}
