% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.tools-package.R
\docType{package}
\name{network.tools-package}
\alias{network.tools}
\alias{network.tools-package}
\title{network.tools: Tools to Analyse and Visualise Bipartite Networks}
\description{
Tools to analyse and visualise bipartite networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Pakillo/network.tools}
  \item Report bugs at \url{https://github.com/Pakillo/network.tools/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rodriguez-Sanchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})

}
\keyword{internal}
