% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{web}
\alias{web}
\title{Bipartite network}
\format{
\subsection{\code{web}}{

A data frame with 32 rows and 3 columns:
\describe{
\item{Plant}{Plant ID}
\item{Animal}{Animal ID}
\item{Visits}{Number of visits}
}
}
}
\source{
dataset simulated
}
\usage{
web
}
\description{
Example data frame containing data for a bipartite web of number of visits
from 8 animals to 4 plants.
}
\keyword{datasets}
