% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide2long.R
\name{wide2long}
\alias{wide2long}
\title{Transform interaction data from wide to long format}
\usage{
wide2long(mat = NULL, int.name = "Visits")
}
\arguments{
\item{mat}{A matrix containing interaction data. Plants in rows, Animals in columns.}

\item{int.name}{character. Column name for the interaction values.}
}
\value{
A data frame with three columns: "Plant", "Animal", and interaction values.
}
\description{
Transform interaction data from wide to long format
}
\examples{
data(web)
mat <- long2wide(web)
mat
df <- wide2long(mat)
df

}
